add binary encoding tests

+++ src/encode/int.rs
@@ -2,6 +2,7 @@ use crate::{encode::Encoder, error::MsgPackErr};
 use std::io::Write;
 
 impl<W: Write> Encoder<W> {
+    #[allow(clippy::cast_possible_truncation)]
     pub(crate) fn encode_i64(&mut self, value: i64) -> Result<(), MsgPackErr> {
         if (0..=127).contains(&value) {
             self.w.write_all(&[value as u8])?;
@@ -29,7 +30,7 @@ impl<W: Write> Encoder<W> {
             self.w.write_all(&(value as u32).to_be_bytes())?;
         } else {
             self.w.write_all(&[0xd3])?;
-            self.w.write_all(&(value as i64).to_be_bytes())?;
+            self.w.write_all(&value.to_be_bytes())?;
         }
 
         Ok(())
